エラーハンドリングの再利用性を考慮
