+++ app/errors/forbidden.rb
@@ -1,0 +1,2 @@
+class Forbidden < StandardError
+end
